==> notebooks/04_time_series_analysis.ipynb <==
# トピックについての時系列を分析する

# 必要なライブラリをインポート
import pandas as pd
import numpy as np
import re
import matplotlib.pyplot as plt
import json

# Jupyter Notebook用の設定
%matplotlib inline

# データを読み込む
data_path = 'data/processed/android_cleaned_mecab_with_topics.csv'

df = pd.read_csv(data_path)

# データの確認
df.head()

# 日付をdatetime型に変換
df['date'] = pd.to_datetime(df['date'])

# 週ごとのレビュー数をカウント
df['week'] = df['date'].dt.to_period('W')

# 週ごとのレビュー数をカウント

weekly_review_counts = df.groupby(['week', 'topic_name']).size().reset_index(name='review_count')

print(weekly_review_counts)

import matplotlib.pyplot as plt
from matplotlib.ticker import MaxNLocator
from matplotlib import rcParams

# フォントをIPAフォントに設定
rcParams['font.family'] = 'Hiragino Sans'

# カラーマップから12色を取得
colors = plt.cm.tab20.colors[:12]

# プロット例
fig, ax = plt.subplots(figsize=(15, 7))

# トピックごとに色を設定
for i, topic_name in enumerate(weekly_review_counts['topic_name'].unique()):
    color = colors[i % len(colors)]  # 12色を循環
    weekly_review_counts[weekly_review_counts['topic_name'] == topic_name].plot(
        x='week', 
        y='review_count', 
        ax=ax, 
        label=topic_name, 
        color=color
    )
    
    
# 縦軸を整数に設定
ax.yaxis.set_major_locator(MaxNLocator(integer=True))

plt.title('週ごとのレビュー数', fontsize=16)  # 日本語タイトル
plt.xlabel('週', fontsize=14)  # 日本語ラベル
plt.ylabel('レビュー数', fontsize=14)  # 日本語ラベル
plt.grid()
plt.legend()
plt.show()


# トピック11を追加する前にリセット
top_topics = pd.Index(['1_広告_れる_表示_多く', '3_fi_wi_wifi_接続', '2_動画_再生_視聴_ポイント', '6_アプリ_こと_表示_停止',
                       '5_クイズ_難しい_難し_難点'])

# トピック11を追加
top_topics = top_topics.append(pd.Index(['11_問題_許可_ページ_ある']))

# トピック-1と0を除外
pivot_df_filtered = pivot_df[top_topics]

# プロット例
fig, ax = plt.subplots(figsize=(15, 7))

# 積み上げ棒グラフを作成
pivot_df_filtered.plot(kind='bar', stacked=True, ax=ax, color=colors[:len(top_topics)])

# 縦軸を整数に設定
ax.yaxis.set_major_locator(MaxNLocator(integer=True))

# 週のラベルを見やすくする
ax.set_xticklabels(pivot_df_filtered.index.strftime('%Y-%m-%d'), rotation=45, ha='right')

plt.title('週ごとのレビュー数', fontsize=16)  # 日本語タイトル
plt.xlabel('週', fontsize=14)  # 日本語ラベル
plt.ylabel('レビュー数', fontsize=14)  # 日本語ラベル
plt.grid()
plt.legend(title='トピック')
plt.show()
